+++ 3rdparty/vulkan/include/CHANGELOG.md
@@ -8,7 +8,7 @@ Other changes:
 
 - Added macro `VMA_DEBUG_DONT_EXCEED_HEAP_SIZE_WITH_ALLOCATION_SIZE` with default value 1.
 - Changed macro `VMA_DEBUG_DONT_EXCEED_MAX_MEMORY_ALLOCATION_COUNT` default value from 0 to 1.
-- Added documentation chapter "Frequenty asked questions".
+- Added documentation chapter "Frequently asked questions".
 - Other fixes and improvements, including compatibility with various platforms and compilers.
 
 # 3.2.1 (2025-02-05)